requirements.txt
Django
google-cloud-bigquery